requirements.txt
alembic==1.17.0
alembic_utils==0.8.8
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
exceptiongroup==1.3.0
fastapi==0.120.0
flupy==1.2.3
h11==0.16.0
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
parse==1.20.2
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
